+++ api.banco_medellin/src/main/java/co/edu/usbbog/bdd/rest/TransaccionController.java
@@ -2,7 +2,6 @@ package co.edu.usbbog.bdd.rest;
 
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,7 +11,6 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import co.edu.usbbog.bdd.model.Transaccion;
 import co.edu.usbbog.bdd.repository.ITransaccion;
 
@@ -22,35 +20,59 @@ public class TransaccionController {
 
 	@Autowired
 	ITransaccion it;
-	
-	@PostMapping("/create") //localhost:3001/ciudad/create
+
+	@PostMapping("/create")
 	public void insertTransaccion(@RequestBody Transaccion c) {
 		it.save(c);
 	}
-	
+
 	@GetMapping("/all")
-	public List<Transaccion>findAllTransacciones() {
-		return it.findAll();
+	public List<Transaccion> findAllTransacciones() {
+		List<Transaccion> l = it.findAll();
+		if (l.isEmpty() || l.equals(null)) {
+			throw new RuntimeException("No hay transacciones registradas");
+		} else {
+			return l;
+		}
 	}
-	
+
 	@GetMapping("/find/{id}")
 	public Optional<Transaccion> findTransaccion(@PathVariable("id") long id) {
-		return it.findById(id);
+		Optional<Transaccion> t = it.findById(id);
+		if (!t.equals(null)) {
+			return t;
+		} else {
+			throw new RuntimeException("Tramsaccion identificada con el ID: " + id + " no encontrado");
+		}
 	}
-	
+
 	@GetMapping("/count")
 	public long coundTransaccion() {
-		return it.count();
+		long c = it.count();
+		if (c != 0) {
+			return c;
+		} else {
+			throw new RuntimeException("No hay transacciones registradas");
+		}
 	}
-	
+
 	@DeleteMapping("/delete/{id}")
 	public void deleteTransaccion(@PathVariable("id") long id) {
-		it.deleteById(id);
+		Optional<Transaccion> t = it.findById(id);
+		if (!t.equals(null)) {
+			it.deleteById(id);
+		} else {
+			throw new RuntimeException("Tramsaccion identificada con el ID: " + id + " no encontrado");
+		}
 	}
-	
+
 	@PutMapping("/update")
 	public void updateTransaccion(@RequestBody Transaccion c) {
-		it.save(c);
+		Optional<Transaccion> t = it.findById(c.getId());
+		if (!t.equals(null)) {
+			it.save(c);
+		} else {
+			throw new RuntimeException("Tramsaccion identificada con el ID: " + c.getId() + " no encontrado");
+		}
 	}
-	
 }
